Se refactoriza el componente Miembros a varios componentes en /componentes/Miembros
